requirements.txt
requests==2.31.0
numpy==1.26.4
pandas==2.1.4
matplotlib==3.8.3
python-dotenv==1.0.1